Trigger GC for actors when they tell the cycle detector they're blocked
Prior to this commit, if an actor blocked, it did not run GC to free
any memory it no longer needed. This would result in blocked actors
holding on to (potentially lots of) memory unnecessarily.

This commit causes GC to be triggered when the cycle detector asks
an actor if it is blocked and the actor responds telling the cycle
detector that it is blocked and also when the actor initially tells
the cycle detector that it is blocked. This should result in memory
being held by blocked actors to be freed more quickly even if the
cycle detector doesn't end up detecting a cycle and reaping the
actors.