Fix race conditions due to work being done after markempty
Prior to this commit, there could be race conditions with a actor
running concurrently on two different threads due to work being
done after an actors queue was marked empty.

This commit reworks the logic to ensure that the actors queue being
marked empty is the last thing that occurs in `ponyint_actor_run`
to prevent these race conditions.

This commit also adds logic in the cycle detector to not send any
messages to actors that are marked as pending destroy to ensure
that the cycle detector cannot create a race condition by sending
messages to a zombie actor (with `rc == 0`) that is about to be
reaped in the near future.