Fix link errors on macOS Big Sur
With the change from Catalina to Big Sur, Apple moved the location of
libraries. On Big Sur, it is required to pass the location of the System
library. The `-lSystem` option is still required for macOS 10.15.7 and
lower.

We don't currently distinguish between different macOS versions, so we
should keep both options. As long as the absolute path of the System
library is present, the `-lSystem` option won't pose any problems on Big
Sur.