Add sending events when nightly releases are finished
Adds sending a GitHub repository dispatch event to ponylang/shared-docker
when all 3 nightly docker images are done building.

This will allow us to trigger nightly shared-docker image rebuilds when
the new images are available rather than hoping that the times line up.

Additional triggered jobs will be added once I know this is working.
For example, triggering testing against the latest ponyc master images
in our various repos.

Additional repos to send to can be added to the `repos` section of the
job's matrix.

At this time, we don't send any payload data with the event as I'm not
aware of what we might need to include.