Change repository_dispatch event sent from cloudsmith event
Previously I had this set to wait on 3 different jobs completing
and then send an event for the nightly being done.

That doesn't work as each of those jobs is triggered by different
incoming events from cloudsmith. There's no way to join on all 3
completing.

Instead I've switched to only sending an event for when the musl
nightly is done as that is all we need at the moment. We can add
more as needed later.