Run standard library tests without the cycle detector
At the moment, there's a race condition that can cause a segfault when
garbage collecting an actor if the cycle detector is on. I haven't had
the solid block of time to fix the problem.

In the meantime, there have been distracting CI failures.

This commit turns off the cycle detector for standard library tests. This
will give us more confidence in those tests. If there is a segfault, we
know it isn't from the known issue.

The changes in this commit will be removed along with the fix for the
race condition.