Turn off zlib compression in LLVM tools
With it on, if we build an environment with libz installed, it ends
up creating unresolved symbols in libponyc that we don't want for
the standalone lib.

We don't need zlib, so we can turn this off. If we need zlib support
in the future, we'll have to figure out another solution like including
zlib.a in libponyc-standalone.a.