Fix compiler crash related to type parameter references
This fixes a bug in rescoping. Rescoping happens after a method was
copied from a trait/interface to another type. It starts with a fresh
scope, visits all of the method's AST nodes and adds the visited symbols
to the method's symbol table. This includes type parameters. If a type
parameter reference is visited, the referenced type parameter is looked
up and added as data to the node.

As rescoping starts with a fresh scope, adding the type parameter to the
reference fails if the type parameter itself is not defined yet. This is
fixed by eagerly adding type parameters to the scope before visiting
type parameter references.

Fixes #3285