Fix early pipe shutdown with Windows ProcessMonitor
This commit fixes a fun bug that results from multiple bits coming together.

First, the ProcessMonitor when running on Windows doesn't using ASIO. Instead
of having events pushed to it, we poll periodically to see if there's anything
to read or if we can write any pending writes.

Because of the polling, it is possible that we will attempt to read when there's
no data in the pipe. When using non-blocking IO with pipes, a read that would return
0 bytes returns an error and sets the error code to a no data error code.

We were interpreting the no data code as it should be when using blocking io. When
using non-blocking io, there error communicates "no data right now, try again later"
which wasn't what we were doing.

The same incorrect handling of the no data error code is also present in the write
code for Windows and is fixed by this commit.

Fixes #3674