Fix literal inference through partial function
Before this change, code such as `1~add(2)` would hit an assertion error
when the compiler tried to infer the type of the literal `2`. The compiler
tries to find the type of the receiver of the function, in this case
`1`, but it lacked the ability to do so when using partial functions. In
those cases, the compiler would try to look at the type of the `~`
token, which is not a valid value literal, and as such it would fail.