Fix memory corruption with Array.chop and String.chop
Both Array.chop and String.chop violate an expectation that the memory
allocator had.

The memory allocator was written before the standard library and assumed
it would be responsible for doing "smart things" with realloc. That is,
when you ask for a realloc, perhaps you don't really need it. Perhaps
you already have enough.

This worked without issue in the standard library until `chop` came along.
Chop turns a single previous object allocation into 2 and that confused
the memory allocator and could lead to it not reallocating (malloc and copy)
when safety required it. Instead it would return the same pointer it was given.

Based on how the Pony standard library ended up, with String and Array understanding
how much they allocated, we do not need that realloc behavior.

This commit removes the behavior from realloc and leaves it so that it will allows
do a malloc and copy which is what the standard library code expects it to do.