Fix type constraint check against NullablePointer being ommitted in FFI declarations
This commit makes the expr pre check for use commands more fine-grained:
we're only interested in ignoring the "guard" part of a use command, but
we still want to go through the arguments to check for any possible
errors during the expr pass. In particular, we're interested in the
NullablePointer constraint check.