Remove whitespace-only lines from stdlib
This small change removes whitespace-only lines from files in the
standard library. All I did was run this with GNU Sed:

    sed -i -sE 's|^\s+$||' packages/**/*.pony

I don't think this merits a changelog update.

There may be some long-term improvements to make here, like an update to
the style guide or adding some sort of autoformatter, but this is my
first contribution!