[ci] use default CPU/Memory
I've checked CPU charts for a few build of `ponylang/ponyc` repo and it seems the tasks are not using all the cores and memory.

Having less CPU request will help to run all tasks faster. Right now you run just 2 at a time and will be able to run 8 at a time.