Don't allow PONYPATH to override standard library
Prior to this change, a library could contain a package called `builtin`
that would override to standard library version. This could be an attack
vector. You can still override the standard library location by using the
ponyc `--paths` option.

Any code which relied on the PONYPATH items being before the standard
library in the package search path will need to switch to using `--paths`
on the ponyc command line.

Closes #3779