Prevent non-opaque structs from being used as behaviour parameters
When an immutable object is sent across actors, the garbage collector
traces its associated object graph, using the type descriptor of the
root object. Structs don't have a type descriptor, so the GC can't trace
them.

An option is to track additional information for objects without type
descriptors, but this incurs in severe overhead for the normal case.

Another option, what this commit does, is to disallow non-opaque objects
without type descriptors to be the root of an object graph when used as
parameters to behaviours. Essentially, this means that one has to wrap
non-tag structs inside another object with a type descriptor. This also
applies to structs inside tuples, since the GC will try to trace each
element in the tuple.