Add support for prebuilt Rocky Linux versions
With CentOS 8 no longer being supported as of the end of 2021, many of
the former CentOS users are migrating to Rocky Linux which offers the
same model that CentOS previously offered.

By adding nightly and release version now, we'll be ready when CentOS 8
is discontinued to provide a path for any CentOS pony users.

Another PR will come once we have a release version of pony out and we
have added support in ponyup to add installation instructions for Rocky
Linux.