Make FileMode._os public
FileMode has always had a private method for getting an integer
representation of the file mode. However, it was private and only
available for use within the `files` package.

I discovered this when writing a pure pony tar library and needed
exactly the functionality of _os.

This commit changes `_os` from private to public as `os` and updates
its callers.