Rename FileMode.os to FileMode.u32
Joe pointed out that the previous (not yet released change) that made
FileMode._os could be better. u32 is a more clear name. _os didn't
really matter as a name as it was private. With the functionality
becoming public we need the best name possible. u32() makes perfect
sense as the name as it is the filemode in a U32 representation.