Clean up child process exits on Windows
In Windows we are closing the process handle when the process successfully exits. However, the wait timer in `ProcessMonitor` repeatedly calls wait on the child, even after it's exited.

This change makes `ProcessMonitor` and `_ProcessWindows` cache final values for the process exit status, so repeated calls to wait on the child will not be referencing an invalid handle.

This also fixes a combinatorial explosion of timers in `ProcessMonitor._wait_for_child()`.