Remove potential flakiness from CodeGen CycleDetector test
By using "sleep" calls, the previous version was locking up a scheduler
thread that could result in the cycle detector not getting a chance to
run. While unlikely to happen given the fairly quick runtime of the test,
it was still possible.

The new version switches from using "sleep" to using a Timer.