Don't remember test "binaries" after running tests
When debugging issues that happen on occassion, if you build the
tests via a command line like `make test-stdlib-debug` and it passes
the first time, you have to go through an entire rebuild and run again
waiting for a failure to have a binary that isn't removed. It can get
very annoying.

I am always removing these lines when testing. I don't think they are
necessary as they are placing items in the build/ directory which can
be tidied up with `make clean`.