Fix non-release build crashes on Arm
The code removed in this commit dates to the very first Arm 32 support
that Sylvan added. It appears during the elapsed years, that the bug
that was being worked around was fixed. The workaround is now itself
a source of odd behavior and crashes for me as I try to get `main`
working on a Raspberry Pi 4 running a 32-bit version of Raspbian.

With the code as it stands before this patch is applied, I get
crashes when running the standard library tests in debug mode. I
can reliably reproduce using the `serialise` tests but it was
also happening "randomly" with other tests as well.

After removing Sylvan's workaround, I have seen a total lack of the
issue.

Closes #3851