Remove old LLVM bug workaround
We can't add the regression test for this as part of automated without
adding `Stringable` to our slimmed down builtin for the JIT tests.

This has been verified by hand and as part of an issue to remove the
JIT tests and replace will running full pony code, we have a note to
add the regression for this.