Remove flakey libponyc test
The CycleDetector test that this commit removes isn't reliable.
Given the correct set of circumstances (64-bit Arm seems to be enough),
it will fail.

I have rewritten it to be "not flakey". That change will come with the
commit that is in progress to remove all JIT tests. Work towards
that goal is available at https://github.com/ponylang/ponyc/pull/3865.

I'm removing this test now as it is one of the last impediments to us
being able to turn on CI for Arm64 tests using Graviton instances in
AWS.

We will have a working version of the test before too long as part
of the remove JIT work and I'm not worried about this getting broken
in the meantime without CI coverage.

If any possible breakage PR comes through before this is again covered
via CI, I can run the in progress JIT-less test version against said
PR.