Add a nightly stress test CI job
Runs message-ubench on 3 platforms making sure we don't get any
crashes. This can help verify that we don't have bugs that
have snuck past the shorter PR tests.

We run both release and debug versions of ubench to check for any
issues with either optimization level.

A debug version of the ponyc runtime is used so that we will trigger
any asserts if they are violated.

Currently the stress test is set to run for 30 minutes for each version.