Reduce the number of x86 Linux PR tasks
We haven't found any value from testing the various different Linux
distributions. It might protect a breakage on a given distro before
it ends up as a failure of a nightly build, but in practice that
hasn't happened.

We now have good alerting on our nightly tasks failing so if it were
to happen, we should be alerted fairly quickly.

Areas where we have seen issues are now covered by these PR tests.

- glibs vs musl
- debug and release runtimes

As part of this PR, we are lowering the memory requested by each
PR task has history has shown that we don't ever use more than 5 gigs.
The 24 gigs we were requesting ended up being overkill.