Refactor pony actor pad size as a function instead of preprocessor value.
From the perspective of LLVM optimizations, this easily-inlined function
that returns a constant value should be the same a preprocessor value.

This is desirable because if the actor pad comes from a function value,
and if the runtime is compiled as LLVM bitcode, it will be possible
for the Pony compiler to change the actor pad by modifying that
function's return value within the LLVM IR (decoded from the bitcode),
prior to linking LLVM optimizations and prior to linking with the user
program.

This means future work will be able to allow features like "memtrack"
(which need to change the size of the actor pad, among other things)
to be toggleable at Pony program compile time, provided that the
Pony user is using the `--runtimebc` feature.