Try to get cirrus failures notifications working as intended
The intent with the cirrus failures workflow is to send alerts
to Zulip for scheduled job failures from Cirrus CI.

The original version I wrote didn't work for all cases because,
the pull_requests field of check_run (and check_suite) isn't filled
out for PRs from forks. The end result of this is that scheduled
job failures AND failure that came from PRs from forks would be
reported.

This commit adds an additional check. If the sender of the event
is "ghost", then don't alert. I added this check as it appears
that all failures coming from CirrusCI for pull requests from
forks have "ghost" as the user.

Hopefully with this additional filtering it will work as expected.
If it does, I'll start updating other  repos as well.