Remove library mode option from ponyc (#3975)
From the very early days of the Pony runtime, a "library mode" has been around that allows you to compile a Pony program to a C compatible library that you can then start up using various runtime APIs to do things like initialize the runtime, create actors, send messages and more. We've made extensive changes to the runtime over the years and have no faith that library mode and its related functionality work.

Our lack of faith extends back many years and we've stated as the Pony core team that we don't consider library mode to be supported; until now, we haven't done anything about what "not supported" means. This commit is the first move in removing support for using the Pony runtime via the various untested C API functions.

Additional changes will be made to take various PONY_API methods that we consider to be "internal only" and rename them to be `ponyint_` to clarify that they aren't intended for calling from external source and to remove PONY_API methods that aren't used internally by the runtime or the standard library.

Closes #3464 
Closes #1882