Add "nodoc" annotation
The can be used to control the pony compilers documentation generation,
any structure that can have a docstring (except packages) can be
annotated with \nodoc\ to prevent documentation from being generated.

This replaces a hack I put in years ago to filter out items that
were "for testing" by looking for Test and _Test at the beginning of
the name as well as providing UnitTest or TestList.

All standard library test items have been annotated with \nodoc\ to
keep them continuing to not have documentation generated by our
standard process of creating the standard library documentation site.

ponylang org libraries will need to be updated as well.

Note that the "should we include this package" hack to filter out
packages called "test" and "builtin_test" still exists for the time
being until we have further discussion.

Closes #3332