Fix bug in #3993
I was storing the type for the wrong object into the object map
in acquire_object. This led to ponyup going boom.

Other programs would as well, but ponyup exhibited.

I'm going to spend some time thinking about how to test this, but
at the moment I don't know the specific combo of Pony code that
got us here.