Change some pure compilation tests into runner tests (#3996)
The code is easier to maintain when its not in C strings so we
decided to transition any that are only "does it compile" tests.

When moving these that previously stopped at a given pass, a decision
was made to add a stub `actor Main` to them. It was either this
or enhancing the runner to be able to only run through a specific pass.

I decided that the `actor Main` stub is better for testing as it
gives us continued coverage that things work even if changes come
in a later pass or the functionality tested moves to a later pass.