Fix compiler crash related to using tuples as a generic constraint
Tuple types aren't legal constraints on generics in Pony and we have a check in
the syntax pass to detect and report that error. However, it was previously
possible to sneak a tuple type as a generic constraint past the syntax check by
smuggling it in a type alias.

Type alias aren't flattened into their various components until the flatten
pass which follows the syntax pass. This means that the following code would
cause a compiler assertion:

```ponyc
type Blocksize is (U8, U32)
class Block[T: Blocksize]
```

This commit adds an additional check in the flatten pass to report an error on
smuggled tuple types as generic constraints.

Fixes #3655