Turn off running debug build Windows tests (#4007)
They are failing fairly regularly and we need to look into why some on the runner tests hang. 

Gordon think that it might be related to other Windows failures we've seen lately (which I think are corral related).
Gordon suspects that all the issues might be tied together by an unknown issue in the Windows Process Monitor code.