Add two while/else tests
While I was recently working on updates to `refer_while`, I was suffering from
a lack of confidence in the test coverage for while/else. The `else` clause isn't
used often in the standard library and I couldn't find tests that explicitly tested
expected while/else functionality (although there are some that seem to depend on it).
I was running tests similar to the two in this commit manually and decided after a bit
of thought that it makes sense to add them to our regular test suite to provide more
confidence going forward.