Fix compiler crash related to using tuples in a union as a generic constraint
Tuple types aren't legal constraints on generics in Pony and we have a check in
an early compiler pass to detect and report that error. However, it was
previously possible to "sneak" a tuple type as a generic constraint past the
earlier check in the compiler by "smuggling" it in a type union in a type
alias.

The type aliases aren't flattened into their various components until after the
code that disallows tuples as generic constraints which would result in the
following code causing ponyc to assert:

```ponyc
type Blocksize is (U64 | (U8, U32))
class Block[T: Blocksize]
```

We've added an additional check to the compiler in a later pass to report an
error on the "smuggled" tuple type as a constraint.

Fixes #4016