Fix the signature of Iter's map_stateful/map to not require ephemerals (#4026)
This fixes a small stdlib issue discovered during: https://github.com/ponylang/ponyc/pull/4018

The signatures for map and map_stateful in itertools were overly strict, requiring `B^` instead of just `B`, but only returned `Iter[B]`. Thus it had to be instantiated with `iso^` in some cases, exercising a bug with double ephemerals.

This is not a breaking change.