Fix failed initialization bug in Process Monitor
Previously, when a process failed to start up, it was possible to get a report
that the process had started up and exited with the error code 0 which
normally signals success.

The problem comes from some rather poor state handling in the `process` package.
This commit doesn't fix the larger issue of all kinds of weirdness around state
handling but it does address the specific bug that was most likely to be
triggered in our unit tests.

The flow in its most extreme form was:

- ProcessMonitor is created in test runner
- Long test is setup
- Dispose when done is setup to send dispose to ProcessMonitor
- ProcessMonitor never successfully starts a Process so its child is _ProcessNone
- notify gets called with `failed`
- `dispose` is sent to the ProcessMonitor
- ProcessMonitor extracts exit code from _ProcessNone which was 0 <-- o boy
- "Success exit code" of 0 is passed as argument to `dispose`