Allow overriding the return type of FFI functions
On the sync call for March 1st, we decided to re-add the ability to specify a
return type for FFI function calls. This was previously removed by RFC 68 on the
grounds that, with FFI declarations being mandatory, specifying a return type
would be redundant.

Before RFC 68 was implemented, the return type at the call site was not allowed
to differ from the return type of the declaration. We have lifted this
limitation, which should help users write FFI bindings for C functions that
might return different types depending on how they are called.

The return type of an FFI function without an explicit return type will default
to the return type of its declaration.